SPOI-8044: Separated out dynamic partitioning into separate Partitioner class
